Remove the surplus debug code from a previous edit.
I came across a problem using the library due to the reader.number not being initialized. In this code it is done coincidentaly by parseCommandLine(). So, in case people copy or leverage this code I've added an explicit reset of the reader number to the default of 0.

To help clean that up we should either add a readerInit() function in the library, or use readerSetOptions, or some way to ensure that reader.number is always determined before calling readerConnect().