Changed polling delay units from microseconds to milliseconds to make command line options passed for this variable compatible between command line and Gtk+ version. Changed timeout units for Gtk+ version from seconds to milliseconds and to sleep the same as interactive version.
Modified code in cpanel.c so that it contains all the GUI stuff, but not start-up stuff (command line parsing, etc), and no main().
Then I broke it up into systemTray, controlPanel, and aboutDialog - each with a source file and header and minimal coupling between them.

This enables systemTray to be called directly from tageventor.c as an optional GUI component.
systemTray in turn can start the controlPanel dialog.
Either systemTray or controlPanel can either cause the AboutDialog to be opened.

Currently using a number of compile flags to allow each of those components to be included or not
These are:
#define BUILD_SYSTEM_TRAY
#define BUILD_CONTROL_PANEL
#define BUILD_ABOUT_DIALOG

The command line version is built using Makefile without any of those defined.
Thus it is functionally the same as the previous versions to all these big changes, and it should also be able to be built on a system without GTK+/GNOME.