Removed the "fakeTags" only used during testing and (for now) always load the defaultCommands.
Initialized the FileChooser widget to chose the folder for script search with the configured folder if it's set.
Added "/etc/tagEventor" folder as a short-cut in the FileChooser dialog to ease choosing folders.

Added a cancel button to control panel window to allow a user to exit directly without saving changes or being asked to save changes.
Due to that, the Close button now has logic and is only sensitive when no change is pending to be saved.

At the moment any changes to the rules made are maintained in memory, and shown again when the control panel is re-opened.
We can debate what the correct behaviour should be....

Made the execution of scripts to be driven by a table of rules, which is loaded (from a constant table at the moment) at start-up.
The rule table is traversed until one is found that matches (todo!) and which the script is found, and executes....