Fixed a bug that meant the GUI version was ignoring a command line setting of the polling delay.
Added code to take setting of settingsDialog() for poll delay and apply it, killing old timeout and
adding new one with correct time. Also sets the slides scale from the initial value, either the
default or that specified on the command line.

Reduced the minimum pollDelay in millsecconds from 1000 to 500

Made changes to tagEventor.desktop to try and keep GNOME happy, as it was leaving messages in syslog, although still not working correctly.