Changes the name of "control panel" to "rules editor" in source file name, function name, comments, variables names etc.
This is prior to cleaning up the whole concept of what was the control panel.
Soon it will have clearly defined concept of "tag event" and an "action" to
be performed when the event occurs.
Combinations of "tag events" and "actions" will be "rules, and the place to
define them will be in the "rules editor"