Extended code to use an array of readers. For now always uses element 0... This required changes to tagReader.c and tagReader.h
Added a new function readerManagerConnect() to allow to connect to the PCSCD daemon once even when multiple readers are being used.
If you use this, be sure to set the hContext element of the tReader struct with what is returned.
The previous method using readerConnect() also will work as it checks the hContext element of the struct and will connect to the reader manager (pcscd) if needed.

Reduced extra stderr output from tagReader.c