Added a file containing string constants for UI and output, will gradually move strings in there over time.
Separated readerDisconnect from readerManagerDisconnect, created that new function.

Added functionality so that readermanagerConnect() enumerates all the readers in the system when called,
and readerConnect() does so also as it may change over time as they are plugged and unplugged.

Made sure that if PCSCD fails, or is not started, that the error is correctly reported on UI and that
it keeps trying to connect to it. Now, if PCSCD is re-started AFTER tagEventor runs, then it will
be detected and connected to.