Centralized all the information about readers and tags under the tReaderManager structure which can be passed easily around, and readers are not part of it.
Parameters that control the behaviour of the application AND the library are in the 
tReaderManager structure and are not unique globals, cleaning up the behaviour of
the library in particular and putting things entirely in control of the application
that should pass appropriate structures to the library.

Added a readersInit() fucntion to make sure that all structures and data are initilized
correctly, and not relying on initialization of global variables by the loader at
start-up. Partly due to the shift mentioned about, but to also allow a potential
re-start from scratch.

I decided I wanted the "driver" code to make as few external (and unpredictable) calls
as possible, and so have removed it's logging code temporarily. I will consider passing
the driver a callback() function it cal use to log detailed information about what it's
doing.