Broke out the detailed description of each tag's logical contents into a separate file, with the first one being tagMifareUltra.h. This removed the need for tagSpecs.h.
Added pointer to a tTag's contents in tagReader.h preparing for when the tag's contents will
get read when it's detected. This is set to NULL in the driver that read's the tags and is passed
to ruleTableEventDispatch, where in the future a new tag's contents maybe read.
This only has a debug trace at the moment.